joysticks: configurable, working, buttons enable
It works!

Press a joystick button to enable that joystick.

Here's why you need to enable the joystick: We don't want your ship to drift
when you have a joystick plugged in and you're not using it, and it doesn't
center perfectly.

You can use the -j/--joystick option to specify which axes you'd like to use
(I've tested it with a PlayStation 3 controller, I can use the tilt sensor to
control VoR like this:

vor -j 4,5

For those with axes but not any buttons (perhaps a handheld computer with an
accelerometer) you can enable a joystick with by adding the joystick number (0
unless you have multiple joysticks) as the third parameter to the -j/--joystick
flag, like this:

vor --joystick=0,1,0

The argument parser is passed the following argv value unless there's a =. When
parsing for your argument, you should barf if you get an invalid value that
looks like it might be a flag (and tell the user they forgot to put in your
argument.) The current parser will handle these forms:

vor -j 0,1
vor --joystick=0,1
vor --joystick 0,1

Still need to make sensitivity adjustable and enable reversing axes.